move CFHT code too
