+++ py/legacyzpts/legacy_zeropoints.py
@@ -210,105 +210,6 @@ def create_annotated_table(T, ann_fn, camera, survey, mp):
 def getrms(x):
     return np.sqrt(np.mean(x**2))
 
-class MegaPrimeMeasurer(object):
-    def __init__(self, *args, **kwargs):
-        self.camera = 'megaprime'
-        super(MegaPrimeMeasurer, self).__init__(*args, **kwargs)
-        self.pixscale = get_pixscale(self.camera)
-
-        # # /global/homes/a/arjundey/idl/pro/observing/decstat.pro
-        ### HACK!!!
-        self.zp0 =  dict(g = 26.610,
-                         r = 26.818,
-                         z = 26.484,
-                         # Totally made up
-                         u = 26.610,
-                         )
-                         #                  # i,Y from DESY1_Stripe82 95th percentiles
-                         #                  i=26.758, Y=25.321) # e/sec
-        self.k_ext = dict(g = 0.17,r = 0.10,z = 0.06,
-                          # Totally made up
-                          u = 0.24)
-        #                   #i, Y totally made up
-        #                   i=0.08, Y=0.06)
-        # --> e/sec
-
-        self.primhdr['WCSCAL'] = 'success'
-        self.goodWcs = True
-
-    def get_extension_list(self, fn, debug=False):
-        if debug:
-            return ['ccd03']
-        hdu = fitsio.FITS(fn)
-        return [hdu[i].get_extname() for i in range(1,len(hdu))]
-
-    def get_ut(self, primhdr):
-        return primhdr['UTC-OBS']
-
-    def get_radec_bore(self, primhdr):
-        return primhdr['RA_DEG'], primhdr['DEC_DEG']
-
-    def ps1_to_observed(self, ps1):
-        # u->g
-        ps1band = dict(u='g').get(self.band, self.band)
-        ps1band_index = ps1cat.ps1band[ps1band]
-        colorterm = self.colorterm_ps1_to_observed(ps1.median, self.band)
-        return ps1.median[:, ps1band_index] + np.clip(colorterm, -1., +1.)
-
-    def get_band(self):
-        band = self.primhdr['FILTER'][0]
-        return band
-
-    def get_gain(self,hdr):
-        return hdr['GAIN']
-
-    def colorterm_ps1_to_observed(self, ps1stars, band):
-        from legacypipe.ps1cat import ps1_to_decam
-        print('HACK -- using DECam color term for CFHT!!')
-        if band == 'u':
-            print('HACK -- using g-band color term for u band!')
-            band = 'g'
-        return ps1_to_decam(ps1stars, band)
-
-    def scale_image(self, img):
-        return img.astype(np.float32)
-
-    def scale_weight(self, img):
-        return img
-
-    def get_wcs(self):
-        ### FIXME -- no distortion solution in here
-        # from astrometry.util.util import Tan
-        # return Tan(self.hdr)
-
-        # "pitcairn" reductions have PV header cards (CTYPE is still RA---TAN)
-        return wcs_pv2sip_hdr(self.hdr)
-
-    def read_weight(self, clip=True, clipThresh=0.01, **kwargs):
-        # Just estimate from image...
-        img,_ = self.read_image()
-        print('Image:', img.shape, img.dtype)
-
-        # Estimate per-pixel noise via Blanton's 5-pixel MAD
-        slice1 = (slice(0,-5,10),slice(0,-5,10))
-        slice2 = (slice(5,None,10),slice(5,None,10))
-        mad = np.median(np.abs(img[slice1] - img[slice2]).ravel())
-        sig1 = 1.4826 * mad / np.sqrt(2.)
-        invvar = (1. / sig1**2)
-        return np.zeros_like(img) + invvar
-
-    def read_bitmask(self):
-        # from legacypipe/cfht.py
-        dqfn = 'cfis/test.mask.0.40.01.fits'
-        if self.slc is not None:
-            mask = fitsio.FITS(dqfn)[self.ext][self.slc]
-        else:
-            mask = fitsio.read(dqfn, ext=self.ext)
-        # This mask is a 16-bit image but has values 0=bad, 1=good.  Flip.
-        return (1 - mask).astype(np.int16)
-
-
-
 def measure_image(img_fn, mp, image_dir='images', run_calibs_only=False,
                   just_imobj=False,
                   survey=None, psfex=True, camera=None, **measureargs):
